Create navbar, also enable viewport scaling
Navbar doesn't contain avatar yet as that hasn't been saved as a cookie

Font-size needs to be set to 0 for parent element to prevent whitespace between elements. It's bizarre but the easiest workaround - otherwise there was a random gap between navbar and background. Font size is set back to the default 16 in style.css

Padding 0 on html/body wasn't doing anything so gone. Padding 10 on div#main doesn't do anything because content is centred so also gone